+++ src/aten/sycl/AbsKernel.cpp
@@ -1,0 +1,35 @@
+#include <ATen/Dispatch.h>
+#include <ATen/OpMathType.h>
+#include <ATen/native/TensorIterator.h>
+#include <c10/core/ScalarType.h>
+
+#include <aten/sycl/Loops.h>
+
+namespace at::native::xpu {
+
+template <typename scalar_t>
+struct AbsFunctor {
+  scalar_t operator()(const scalar_t a) const {
+    return std::abs(a);
+  }
+};
+
+void abs_kernel(TensorIteratorBase& iter) {
+  auto dtype = iter.dtype();
+  if (at::isComplexType(dtype)) {
+    AT_DISPATCH_COMPLEX_TYPES_AND(kComplexHalf, dtype, "abs_xpu", [&]() {
+      using opmath_t = at::opmath_type<scalar_t>;
+      gpu_kernel(iter, AbsFunctor<opmath_t>());
+    });
+  } else {
+    AT_DISPATCH_ALL_TYPES_AND3(
+        ScalarType::Half,
+        ScalarType::BFloat16,
+        ScalarType::Bool,
+        iter.dtype(),
+        "abs_xpu",
+        [&]() { gpu_kernel(iter, AbsFunctor<scalar_t>()); });
+  }
+}
+
+} // namespace at::native::xpu
