+++ src/aten/sycl/UnarySignKernels.cpp
@@ -54,4 +54,25 @@ void sign_kernel(TensorIteratorBase& iter) {
   }
 }
 
+template <typename scalar_t>
+struct NegFunctor {
+  scalar_t operator()(scalar_t a) const {
+    return -a;
+  }
+};
+
+void neg_kernel(TensorIteratorBase& iter) {
+  auto dtype = iter.dtype();
+  if (at::isComplexType(dtype)) {
+    AT_DISPATCH_COMPLEX_TYPES_AND(kComplexHalf, dtype, "neg_xpu", [&]() {
+      gpu_kernel(iter, NegFunctor<scalar_t>());
+    });
+  } else {
+    AT_DISPATCH_ALL_TYPES_AND2(
+        ScalarType::Half, ScalarType::BFloat16, dtype, "neg_xpu", [&]() {
+          gpu_kernel(iter, NegFunctor<scalar_t>());
+        });
+  }
+}
+
 } // namespace at::native::xpu
