+++ src/aten/sycl/UnarySignKernels.h
@@ -4,6 +4,8 @@
 
 namespace at::native::xpu {
 
+void neg_kernel(TensorIteratorBase& iter);
+
 void sgn_kernel(TensorIteratorBase& iter);
 
 void sign_kernel(TensorIteratorBase& iter);
