+++ src/aten/sycl/UnarySpecialOpsKernels.h
@@ -6,4 +6,8 @@ namespace at::native::xpu {
 
 void sigmoid_kernel(TensorIteratorBase& iter);
 
+void erf_kernel(TensorIteratorBase& iter);
+
+void erfc_kernel(TensorIteratorBase& iter);
+
 } // namespace at::native::xpu
